+++ estacionamento-api/src/main/java/com/fundatec/ti20/estacionamento/service/TarifaService.java
@@ -1,8 +1,10 @@
 package com.fundatec.ti20.estacionamento.service;
 
+import com.fundatec.ti20.estacionamento.exceptions.ConflitoException;
+import com.fundatec.ti20.estacionamento.exceptions.ObjectNotFoundException;
 import com.fundatec.ti20.estacionamento.model.Conta;
 import com.fundatec.ti20.estacionamento.model.Tarifa;
-import com.fundatec.ti20.estacionamento.model.Veiculo;
+import com.fundatec.ti20.estacionamento.model.enums.TipoTarifa;
 import com.fundatec.ti20.estacionamento.model.enums.TipoVeiculo;
 import com.fundatec.ti20.estacionamento.repository.TarifaRepository;
 import lombok.AllArgsConstructor;
@@ -10,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
+import java.util.List;
 
 
 @Service
@@ -20,30 +23,45 @@ public class TarifaService {
     private final TarifaRepository repository;
 
 
-    @Autowired
-    private final CalcularContaService service;
+    public Tarifa find(TipoTarifa tipoTarifa, TipoVeiculo tipoVeiculo) {
+        return repository.findByTipoTarifaAndTipoVeiculo(tipoTarifa,
+                tipoVeiculo).orElseThrow(() -> new ObjectNotFoundException("Tarifa"));
+    }
 
-//    public Tarifa find(Integer idVeiculo, Integer idConta) {
-//        return repository.findByIdVeiculoAndIdConta(idVeiculo, idConta).orElseThrow(() -> new RuntimeException("Tarifa não encontrada"));
-//    }
+    public List<Tarifa> findAll() {
+        if(repository.findAll().isEmpty()){
+            throw new ObjectNotFoundException("nenhuma tarifa, o banco de dados está vazio");
+        }
+        return repository.findAll();
+    }
+
+    public Tarifa findById(Integer id) {
+        return repository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Tarifa"));
+    }
 
     public Tarifa salvar(Tarifa tarifa) {
+        if (repository.findAll().contains(tarifa)){
+            throw new ConflitoException("tarifa");
+        }
         return repository.save(tarifa);
     }
 
     public void delete(Integer id) {
+        if (!repository.findAll().contains(id)){
+            throw new ObjectNotFoundException("tarifa");
+        }
         repository.deleteById(id);
     }
 
-   public void fecharTarifa(Conta conta) {
-        Veiculo veiculo = conta.getVeiculo();
-        TipoVeiculo tipoVeiculo = veiculo.getTipoVeiculo();
-        if (conta.getVeiculo().temAssinante()) {
-            double valorTarifa = (service.calcular(tipoVeiculo, conta.descobrirDuracaoEmMinutos())) * 0.90;
-        }
-        double valorTarifa = service.calcular(tipoVeiculo, conta.descobrirDuracaoEmMinutos());
-        Tarifa tarifa = Tarifa.builder().valor(new BigDecimal(valorTarifa)).build();
-        salvar(tarifa);
-    }
+    //Revisar método requerido pelo professor
+//    public void MontarTarifa(Tarifa tarifa) {
+//        Conta contaUtilizada = tarifa.getConta();
+//        long tempoEmMinutos = contaService.descobrirDuracaoEmMinutos(contaUtilizada);
+//        TipoVeiculo tipoVeiculo = contaUtilizada.getVeiculo().getTipoVeiculo();
+//        tarifa.setValor(BigDecimal.valueOf(calculoService.calcular(tipoVeiculo, tempoEmMinutos)));
+//        contaUtilizada.setValorPagamento(tarifa.getValor());
+//    }
+
+
 
 }
